Fixed endgame condition
If a player is going at snail speed and a fruit spawned right in
front of the player at a particular frame, fruit accounting may be off
such that player is not able to enter the victory state despite having
ate all the fruit.  This change is meant to fix that by doing an
exhaustive check for endgame after failed to spawn fruit.